Created ScaleDisplay view, better unit switching, speration of logic, etc

+++ Gravity/CoinHolder.m
@@ -58,7 +58,7 @@ static const CGFloat coinSpacing = 20;
 }
 
 - (void) coinSelected:(NSUInteger)coinIndex {
-    NSLog(@"Coin Selected: %ld", coinIndex);
+    NSLog(@"Coin Selected: %d", (int)coinIndex);
     [self.coinSelectionDelegate coinSelected:coinIndex];
     self.activeCoinButtonIndex++;
 }
